requirements.txt
django==4.1.2
djangorestframework
drf-extra-fields
pillow